fix: Fixed naming of Blog class and BlogContext merge leftovers
