+++ src/Modules/Blog/Explorer.Blog.Infrastructure/BlogStartup.cs
@@ -9,39 +9,27 @@ using Explorer.BuildingBlocks.Core.UseCases;
 using Explorer.BuildingBlocks.Infrastructure.Database;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
-using domainBlog = Explorer.Blog.Core.Domain.Blog;
 
-namespace Explorer.Blog.Infrastructure
-{
-    public static class BlogStartup
-    {
-        services.AddScoped<IBlogService, BlogService>();
-        services.AddScoped<IClubService, ClubService>();
-    }
-
-    private static void SetupInfrastructure(IServiceCollection services)
-    {
-        services.AddScoped(typeof(ICrudRepository<domainBlog>), typeof(CrudDatabaseRepository<domainBlog, BlogContext>));
-        services.AddScoped(typeof(ICrudRepository<BlogImage>), typeof(CrudDatabaseRepository<BlogImage, BlogContext>));
-        services.AddScoped(typeof(ICrudRepository<Club>), typeof(CrudDatabaseRepository<Club, BlogContext>));
-
-        public static IServiceCollection ConfigureBlogModule(this IServiceCollection services)
-        {
+namespace Explorer.Blog.Infrastructure {
+    public static class BlogStartup {
+        public static IServiceCollection ConfigureBlogModule(this IServiceCollection services) {
             services.AddAutoMapper(typeof(BlogProfile).Assembly);
             SetupCore(services);
             SetupInfrastructure(services);
             return services;
         }
 
-        private static void SetupCore(IServiceCollection services)
-        {
+        private static void SetupCore(IServiceCollection services) {
+            services.AddScoped(typeof(ICrudRepository<BlogPost>), typeof(CrudDatabaseRepository<BlogPost, BlogContext>));
+            services.AddScoped<IBlogService, BlogService>();
+            services.AddScoped(typeof(ICrudRepository<BlogImage>), typeof(CrudDatabaseRepository<BlogImage, BlogContext>));
             services.AddScoped<IClubService, ClubService>();
             services.AddScoped<IBlogCommentService, BlogCommentService>();
             services.AddScoped<ICommentRepository, CommentRepository>();
+
         }
 
-        private static void SetupInfrastructure(IServiceCollection services)
-        {
+        private static void SetupInfrastructure(IServiceCollection services) {
             services.AddScoped(typeof(ICrudRepository<Club>), typeof(CrudDatabaseRepository<Club, BlogContext>));
             services.AddScoped(typeof(ICrudRepository<BlogComment>), typeof(CrudDatabaseRepository<BlogComment, BlogContext>));
 
@@ -51,3 +39,4 @@ namespace Explorer.Blog.Infrastructure
         }
     }
 }
+
